+++ src/maicos/modules/dielectriccylinder.py
@@ -16,8 +16,6 @@ import scipy.constants
 from ..core import CylinderBase
 from ..lib.util import charge_neutral, citation_reminder, get_compound, render_docs
 
-logger = logging.getLogger(__name__)
-
 
 @render_docs
 @charge_neutral(filter="error")
@@ -91,7 +89,7 @@ class DielectricCylinder(CylinderBase):
         _, self.inverse_ix = np.unique(ix, return_inverse=True)
 
         if zmin is not None or zmax is not None or rmin != 0 or rmax is not None:
-            logger.warning(
+            logging.warning(
                 "Setting `rmin` and `rmax` (as well as `zmin` and `zmax`) might cut "
                 "off molecules. This will lead to severe artifacts in the dielectric "
                 "profiles."
@@ -119,7 +117,7 @@ class DielectricCylinder(CylinderBase):
 
     def _prepare(self) -> None:
         # Print Philip Loche citation
-        logger.info(citation_reminder("10.1021/acs.jpcb.9b09269"))
+        logging.info(citation_reminder("10.1021/acs.jpcb.9b09269"))
 
         super()._prepare()